backend/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
neo4j>=5.14.1
openai>=1.10.0
python-dotenv>=1.0.0
tiktoken>=0.5.2
pydantic>=2.5.3
numpy>=1.24.3